Final Commit
Ready for Review. Renamed the files to their proper names. Added the script to the html. Runs a function when the page loads. Created a function to switch out the style sheets. Randomly selects a style when the page is loaded. Adds in buttons to change the style sheet. Added in comments to the javascript file.